ui: new playlist sidebar
