+++ iina/PlaylistViewController.swift
@@ -8,1040 +8,40 @@
 
 import Cocoa
 
-fileprivate let PrefixMinLength = 7
-fileprivate let FilenameMinLength = 12
+extension SidebarViewController.TabType {
+  static let playlist = SidebarViewController.TabType(0, "playlist", .sf("music.note.list")!)
+  static let chapters = SidebarViewController.TabType(1, "chapters", .sf("list.and.film")!)
+}
 
-fileprivate let MenuItemTagCut = 601
-fileprivate let MenuItemTagCopy = 602
-fileprivate let MenuItemTagPaste = 603
-fileprivate let MenuItemTagDelete = 604
 
-class PlaylistViewController: NSViewController, NSTableViewDataSource, NSTableViewDelegate, NSMenuDelegate, SidebarViewController, NSMenuItemValidation {
+class PlaylistViewController: SidebarViewController {
+  var isInMiniPlayer: Bool = false
 
-  override var nibName: NSNib.Name {
-    return NSNib.Name("PlaylistViewController")
+  override var sidebarType: SidebarController.ViewType {
+    .playlist
+  }
+  override var leadingPrefKey: Preference.Key {
+    .sidebarPlaylistDisplayAtLeading
   }
 
-  init(mainWindow: MainWindowController) {
-    self.mainWindow = mainWindow
-    self.player = mainWindow.player
-    super.init(nibName: nil, bundle: nil)
+  override func allTabs() -> [SidebarViewController.TabType] {
+    [.playlist, .chapters]
+  }
+  override var defaultTab: SidebarViewController.TabType {
+    .playlist
   }
 
-  required init?(coder: NSCoder) {
-    fatalError("init(coder:) has not been implemented")
-  }
-
-  unowned let mainWindow: MainWindowController
-  unowned let player: PlayerCore
-
-  /** Similar to the one in `QuickSettingViewController`.
-   Since IBOutlet is `nil` when the view is not loaded at first time,
-   use this variable to cache which tab it need to switch to when the
-   view is ready. The value will be handled after loaded.
-   */
-  private var pendingSwitchRequest: TabViewType?
-
-  var playlistChangeObserver: NSObjectProtocol?
-
-  /** Enum for tab switching */
-  enum TabViewType: Int {
-    case playlist = 0
-    case chapters
-
-    init?(name: String) {
-      switch name {
-      case "playlist":
-        self = .playlist
-      case "chapters":
-        self = .chapters
-      default:
-        return nil
-      }
+  override func getTabView(for tab: SidebarViewController.TabType) -> SidebarPane {
+    return switch tab {
+    case .playlist: SidebarPlaylistPane(player: player)
+    case .chapters: SidebarChaptersPane(player: player)
+    default: fatalError()
     }
-  }
-
-  var currentTab: TabViewType = .playlist
-
-  @IBOutlet weak var playlistTableView: NSTableView!
-  @IBOutlet weak var chapterTableView: NSTableView!
-  @IBOutlet weak var playlistBtn: NSButton!
-  @IBOutlet weak var chaptersBtn: NSButton!
-  @IBOutlet weak var tabView: NSTabView!
-  @IBOutlet weak var buttonTopConstraint: NSLayoutConstraint!
-  @IBOutlet weak var tabHeightConstraint: NSLayoutConstraint!
-  @IBOutlet weak var deleteBtn: NSButton!
-  @IBOutlet weak var loopBtn: NSButton!
-  @IBOutlet weak var shuffleBtn: NSButton!
-  @IBOutlet weak var sortBtn: NSButton!
-  @IBOutlet weak var totalLengthLabel: NSTextField!
-  @IBOutlet var subPopover: NSPopover!
-  @IBOutlet var addFileMenu: NSMenu!
-  @IBOutlet weak var addBtn: NSButton!
-  @IBOutlet weak var removeBtn: NSButton!
-  
-  @Atomic private var playlistTotalLengthIsReady = false
-  @Atomic private var playlistTotalLength: Double? = nil
-
-  var downShift: CGFloat = 0 {
-    didSet {
-      buttonTopConstraint.constant = downShift
-    }
-  }
-
-  var useCompactTabHeight = false {
-    didSet {
-      tabHeightConstraint.constant = useCompactTabHeight ? 32 : 48
-    }
-  }
-
-  override func viewDidLoad() {
-    super.viewDidLoad()
-    withAllTableViews { (view) in
-      view.dataSource = self
-    }
-    playlistTableView.menu?.delegate = self
-
-    [deleteBtn, loopBtn, shuffleBtn].forEach {
-      $0?.image?.isTemplate = true
-      $0?.alternateImage?.isTemplate = true
-    }
-
-    deleteBtn.toolTip = NSLocalizedString("mini_player.delete", comment: "delete")
-    loopBtn.toolTip = NSLocalizedString("mini_player.loop", comment: "loop")
-    shuffleBtn.toolTip = NSLocalizedString("mini_player.shuffle", comment: "shuffle")
-    addBtn.toolTip = NSLocalizedString("mini_player.add", comment: "add")
-    removeBtn.toolTip = NSLocalizedString("mini_player.remove", comment: "remove")
-    sortBtn.toolTip = NSLocalizedString("mini_player.sort", comment: "sort")
-
-    hideTotalLength()
-
-    // colors
-    withAllTableViews { $0.backgroundColor = NSColor.sidebarTableBackground }
-
-    // handle pending switch tab request
-    if pendingSwitchRequest != nil {
-      switchToTab(pendingSwitchRequest!)
-      pendingSwitchRequest = nil
-    } else {
-      // Initial display: need to draw highlight for currentTab
-      updateTabButtons(activeTab: currentTab)
-    }
-
-    // notifications
-    playlistChangeObserver = NotificationCenter.default.addObserver(forName: .iinaPlaylistChanged, object: player, queue: OperationQueue.main) { [unowned self] _ in
-      self.playlistTotalLengthIsReady = false
-      self.reloadData(playlist: true, chapters: false)
-    }
-
-    // register for double click action
-    let action = #selector(performDoubleAction(sender:))
-    playlistTableView.doubleAction = action
-    playlistTableView.target = self
-    chapterTableView.doubleAction = action
-    chapterTableView.target = self
-
-    // register for drag and drop
-    playlistTableView.registerForDraggedTypes([.iinaPlaylistItem, .nsFilenames, .nsURL, .string])
-
-    (subPopover.contentViewController as! SubPopoverViewController).player = player
-    if let popoverView = subPopover.contentViewController?.view,
-      popoverView.trackingAreas.isEmpty {
-      popoverView.addTrackingArea(NSTrackingArea(rect: popoverView.bounds,
-                                                 options: [.activeAlways, .inVisibleRect, .mouseEnteredAndExited, .mouseMoved],
-                                                 owner: mainWindow, userInfo: ["obj": 0]))
-    }
-  }
-
-  override func viewDidAppear() {
-    reloadData(playlist: true, chapters: true)
-    updateLoopBtnStatus()
-  }
-
-  deinit {
-    NotificationCenter.default.removeObserver(self.playlistChangeObserver!)
-  }
-
-  func reloadData(playlist: Bool, chapters: Bool) {
-    guard player.info.state.active else { return }
-    if playlist {
-      player.getPlaylist()
-      playlistTableView.reloadData()
-    }
-    if chapters {
-      chapterTableView.reloadData()
-    }
-  }
-
-  private func showTotalLength() {
-    guard let playlistTotalLength = playlistTotalLength, playlistTotalLengthIsReady else { return }
-    totalLengthLabel.isHidden = false
-    if playlistTableView.numberOfSelectedRows > 0 {
-      let info = player.info
-      let selectedDuration = info.calculateTotalDuration(playlistTableView.selectedRowIndexes)
-      totalLengthLabel.stringValue = String(format: NSLocalizedString("playlist.total_length_with_selected", comment: "%@ of %@ selected"),
-                                            VideoTime(selectedDuration).stringRepresentation,
-                                            VideoTime(playlistTotalLength).stringRepresentation)
-    } else {
-      totalLengthLabel.stringValue = String(format: NSLocalizedString("playlist.total_length", comment: "%@ in total"),
-                                            VideoTime(playlistTotalLength).stringRepresentation)
-    }
-  }
-
-  private func hideTotalLength() {
-    totalLengthLabel.isHidden = true
-  }
-
-  private func refreshTotalLength() {
-    let totalDuration: Double? = player.info.calculateTotalDuration()
-    if let duration = totalDuration {
-      playlistTotalLengthIsReady = true
-      playlistTotalLength = duration
-      DispatchQueue.main.async {
-        self.showTotalLength()
-      }
-    } else {
-      DispatchQueue.main.async {
-        self.hideTotalLength()
-      }
-    }
-  }
-
-  func updateLoopBtnStatus() {
-    guard isViewLoaded else { return }
-    let loopMode = player.getLoopMode()
-    switch loopMode {
-    case .off:  loopBtn.state = .off
-    case .file: loopBtn.state = .on
-    default:    loopBtn.state = .mixed
-    }
-    loopBtn.alternateImage = NSImage.init(named: loopBtn.state == .on ? "loop_file" : "loop_dark")
-  }
-
-  // MARK: - Tab switching
-
-  /** Switch tab (call from other objects) */
-  func pleaseSwitchToTab(_ tab: TabViewType) {
-    if isViewLoaded {
-      switchToTab(tab)
-    } else {
-      // cache the request
-      pendingSwitchRequest = tab
-    }
-  }
-
-  /** Switch tab (for internal call) */
-  private func switchToTab(_ tab: TabViewType) {
-    updateTabButtons(activeTab: tab)
-    switch tab {
-    case .playlist:
-      tabView.selectTabViewItem(at: 0)
-    case .chapters:
-      tabView.selectTabViewItem(at: 1)
-    }
-
-    currentTab = tab
-  }
-
-  // Updates display of all tabs buttons to indicate that the given tab is active and the rest are not
-  private func updateTabButtons(activeTab: TabViewType) {
-    switch activeTab {
-    case .playlist:
-      updateTabActiveStatus(for: playlistBtn, isActive: true)
-      updateTabActiveStatus(for: chaptersBtn, isActive: false)
-    case .chapters:
-      updateTabActiveStatus(for: playlistBtn, isActive: false)
-      updateTabActiveStatus(for: chaptersBtn, isActive: true)
-    }
-  }
-
-  private func updateTabActiveStatus(for btn: NSButton, isActive: Bool) {
-    btn.contentTintColor = isActive ? NSColor.sidebarTabTintActive : NSColor.sidebarTabTint
-  }
-
-  // MARK: - NSTableViewDataSource
-
-  func numberOfRows(in tableView: NSTableView) -> Int {
-    if tableView == playlistTableView {
-      return player.info.$playlist.withLock { $0.count }
-    } else if tableView == chapterTableView {
-      return player.info.chapters.count
-    } else {
-      return 0
-    }
-  }
-
-  // MARK: - Drag and Drop
-
-  func copyToPasteboard(_ tableView: NSTableView, writeRowsWith rowIndexes: IndexSet, to pboard: NSPasteboard) {
-    do {
-      let indexesData = try NSKeyedArchiver.archivedData(withRootObject: rowIndexes, requiringSecureCoding: true)
-      let filePaths = player.info.$playlist.withLock { playlist in
-        rowIndexes.map { playlist[$0].filename }
-      }
-      pboard.declareTypes([.iinaPlaylistItem, .nsFilenames], owner: tableView)
-      pboard.setData(indexesData, forType: .iinaPlaylistItem)
-      pboard.setPropertyList(filePaths, forType: .nsFilenames)
-    } catch {
-      // Internal error, archivedData should not fail.
-      Logger.log("Failed to copy from playlist to pasteboard: \(error)", level: .error,
-                 subsystem: player.subsystem)
-    }
-  }
-
-  @discardableResult
-  func pasteFromPasteboard(row: Int, from pboard: NSPasteboard) -> Bool {
-    if let paths = pboard.propertyList(forType: .nsFilenames) as? [String] {
-      let playableFiles = Utility.resolveURLs(player.getPlayableFiles(in: paths.map {
-        $0.hasPrefix("/") ? URL(fileURLWithPath: $0) : URL(string: $0)!
-      }))
-      if playableFiles.count == 0 {
-        return false
-      }
-      player.addToPlaylist(paths: playableFiles.map { $0.isFileURL ? $0.path : $0.absoluteString }, at: row)
-    } else if let urls = pboard.propertyList(forType: .nsURL) as? [String] {
-      player.addToPlaylist(paths: urls, at: row)
-    } else if let droppedString = pboard.string(forType: .string), Regex.url.matches(droppedString) {
-      player.addToPlaylist(paths: [droppedString], at: row)
-    } else {
-      return false
-    }
-    player.postNotification(.iinaPlaylistChanged)
-    return true
-  }
-
-  func tableView(_ tableView: NSTableView, writeRowsWith rowIndexes: IndexSet, to pboard: NSPasteboard) -> Bool {
-    if tableView == playlistTableView {
-      copyToPasteboard(tableView, writeRowsWith: rowIndexes, to: pboard)
-      return true
-    }
-    return false
-  }
-
-
-  func tableView(_ tableView: NSTableView, validateDrop info: NSDraggingInfo, proposedRow row: Int, proposedDropOperation dropOperation: NSTableView.DropOperation) -> NSDragOperation {
-    playlistTableView.setDropRow(row, dropOperation: .above)
-    if info.draggingSource as? NSTableView === tableView {
-      return .move
-    }
-    return player.acceptFromPasteboard(info, isPlaylist: true)
-  }
-
-  func tableView(_ tableView: NSTableView, acceptDrop info: NSDraggingInfo, row: Int, dropOperation: NSTableView.DropOperation) -> Bool {
-    if info.draggingSource as? NSTableView === tableView,
-      let rowData = info.draggingPasteboard.data(forType: .iinaPlaylistItem),
-      let indexSet = try? NSKeyedUnarchiver.unarchivedObject(ofClass: NSIndexSet.self, from: rowData) as? IndexSet {
-      // Drag & drop within playlistTableView
-      var oldIndexOffset = 0, newIndexOffset = 0
-      for oldIndex in indexSet {
-        if oldIndex < row {
-          player.playlistMove(oldIndex + oldIndexOffset, to: row)
-          oldIndexOffset -= 1
-        } else {
-          player.playlistMove(oldIndex, to: row + newIndexOffset)
-          newIndexOffset += 1
-        }
-        Logger.log("Playlist Drag & Drop from \(oldIndex) to \(row)", subsystem: player.subsystem)
-      }
-      player.postNotification(.iinaPlaylistChanged)
-      return true
-    }
-    // Otherwise, could be copy/cut & paste within playlistTableView
-    return pasteFromPasteboard(row: row, from: info.draggingPasteboard)
-  }
-
-  // MARK: - Edit Menu Support
-
-  func validateMenuItem(_ menuItem: NSMenuItem) -> Bool {
-    if currentTab == .playlist {
-      switch menuItem.tag {
-      case MenuItemTagCut, MenuItemTagCopy, MenuItemTagDelete:
-        return playlistTableView.selectedRow != -1
-      case MenuItemTagPaste:
-        return NSPasteboard.general.types?.contains(.nsFilenames) ?? false
-      default:
-        break
-      }
-    }
-    return menuItem.isEnabled
-  }
-
-  @objc func copy(_ sender: NSMenuItem) {
-    copyToPasteboard(playlistTableView, writeRowsWith: playlistTableView.selectedRowIndexes, to: .general)
-  }
-
-  @objc func cut(_ sender: NSMenuItem) {
-    copy(sender)
-    delete(sender)
-  }
-
-  @objc func paste(_ sender: NSMenuItem) {
-    let dest = playlistTableView.selectedRowIndexes.first ?? 0
-    pasteFromPasteboard(row: dest, from: .general)
-  }
-
-
-  @objc func delete(_ sender: NSMenuItem) {
-    player.playlistRemove(playlistTableView.selectedRowIndexes)
-  }
-
-  // MARK: - private methods
-
-  private func withAllTableViews(_ block: (NSTableView) -> Void) {
-    block(playlistTableView)
-    block(chapterTableView)
-  }
-
-  // MARK: - IBActions
-
-  @IBAction func addToPlaylistBtnAction(_ sender: NSButton) {
-    addFileMenu.popUp(positioning: nil, at: .zero, in: sender)
-  }
-
-  @IBAction func removeBtnAction(_ sender: NSButton) {
-    player.playlistRemove(playlistTableView.selectedRowIndexes)
-  }
-
-  @IBAction func addFileAction(_ sender: AnyObject) {
-    Utility.quickMultipleOpenPanel(title: "Add to playlist", canChooseDir: true) { urls in
-      let playableFiles = self.player.getPlayableFiles(in: urls)
-      if playableFiles.count != 0 {
-        self.player.addToPlaylist(paths: playableFiles.map { $0.path },
-                                  at: self.player.info.$playlist.withLock { $0.count })
-        self.player.mainWindow.sidebars.playlistView.reloadData(playlist: true, chapters: false)
-        self.player.sendOSD(.addToPlaylist(playableFiles.count))
-      }
-    }
-  }
-
-  @IBAction func addURLAction(_ sender: AnyObject) {
-    Utility.quickPromptPanel("add_url") { url in
-      if Regex.url.matches(url) {
-        self.player.appendToPlaylist(url)
-        self.player.mainWindow.sidebars.playlistView.reloadData(playlist: true, chapters: false)
-        self.player.sendOSD(.addToPlaylist(1))
-      } else {
-        Utility.showAlert("wrong_url_format")
-      }
-    }
-  }
-
-  @IBAction func clearPlaylistBtnAction(_ sender: AnyObject) {
-    player.clearPlaylist()
-    player.sendOSD(.clearPlaylist)
-  }
-
-  @IBAction func playlistBtnAction(_ sender: AnyObject) {
-    reloadData(playlist: true, chapters: false)
-    switchToTab(.playlist)
-  }
-
-  @IBAction func chaptersBtnAction(_ sender: AnyObject) {
-    reloadData(playlist: false, chapters: true)
-    switchToTab(.chapters)
-  }
-
-  @IBAction func loopBtnAction(_ sender: NSButton) {
-    player.nextLoopMode()
-  }
-
-  @IBAction func shuffleBtnAction(_ sender: AnyObject) {
-    player.toggleShuffle()
-  }
-
-
-  @objc func performDoubleAction(sender: AnyObject) {
-    guard let tv = sender as? NSTableView, tv.numberOfSelectedRows > 0 else { return }
-    if tv == playlistTableView {
-      player.playFileInPlaylist(tv.selectedRow)
-    } else {
-      let index = tv.selectedRow
-      player.playChapter(index)
-    }
-    tv.deselectAll(self)
-    tv.reloadData()
   }
 
   @IBAction func prefixBtnAction(_ sender: PlaylistPrefixButton) {
     sender.isFolded = !sender.isFolded
   }
 
-  @IBAction func subBtnAction(_ sender: NSButton) {
-    let row = playlistTableView.row(for: sender)
-    guard let vc = subPopover.contentViewController as? SubPopoverViewController else { return }
-    vc.filePath = player.info.$playlist.withLock { $0[row].filename }
-    vc.tableView.reloadData()
-    vc.heightConstraint.constant = (vc.tableView.rowHeight + vc.tableView.intercellSpacing.height) * CGFloat(vc.tableView.numberOfRows)
-    subPopover.show(relativeTo: sender.bounds, of: sender, preferredEdge: .maxY)
-  }
-
-  @IBAction func sortingBtnAction(_ sender: NSButton) {
-    let menu = NSMenu()
-    if #available(macOS 14.0, *) {
-      menu.addItem(.sectionHeader(title: NSLocalizedString("playlist.sorting.header", comment: "Sorting")))
-    }
-    menu.addItem(withTitle: NSLocalizedString("playlist.sorting.filename_ascending", comment: "Filename Ascending"), action: #selector(sortPathAscending), keyEquivalent: "")
-    menu.addItem(withTitle: NSLocalizedString("playlist.sorting.filename_descending", comment: "Filename Descending"), action: #selector(sortPathDesecnding), keyEquivalent: "")
-    menu.addItem(withTitle: NSLocalizedString("playlist.sorting.path_ascending", comment: "File Path Ascending"), action: #selector(sortPathAscending), keyEquivalent: "")
-    menu.addItem(withTitle: NSLocalizedString("playlist.sorting.path_descending", comment: "File Path Descending"), action: #selector(sortPathDesecnding), keyEquivalent: "")
-    NSMenu.popUpContextMenu(menu, with: NSApplication.shared.currentEvent!, for: sender)
-  }
-
-  @objc func sortNameAscending() { sortName(ascending: true) }
-  @objc func sortNameDesecnding() { sortName(ascending: false) }
-  @objc func sortPathAscending() { sortPath(ascending: true) }
-  @objc func sortPathDesecnding() { sortPath(ascending: false) }
-
-  private func sortName(ascending: Bool) {
-    var playlist = player.info.playlist
-    playlist.sort(by: {
-      let results = $0.filenameForDisplay < $1.filenameForDisplay
-      return ascending ? results : !results
-    })
-    player.playlistReorder(newPlaylist: playlist)
-  }
-
-  private func sortPath(ascending: Bool) {
-    var playlist = player.info.playlist
-    playlist.sort(by: {
-      let results = $0.filename < $1.filename
-      return ascending ? results : !results
-    })
-    player.playlistReorder(newPlaylist: playlist)
-  }
-
-  // MARK: - Table delegates
-
-  func tableViewSelectionDidChange(_ notification: Notification) {
-    let tv = notification.object as! NSTableView
-    if tv == playlistTableView {
-      showTotalLength()
-      return
-    }
-  }
-
-  func tableView(_ tableView: NSTableView, viewFor tableColumn: NSTableColumn?, row: Int) -> NSView? {
-    guard let identifier = tableColumn?.identifier else { return nil }
-    let info = player.info
-    let v = tableView.makeView(withIdentifier: identifier, owner: self) as! NSTableCellView
-
-    // playlist
-    if tableView == playlistTableView {
-      let item: MPVPlaylistItem? = info.$playlist.withLock { playlist in
-        guard row < playlist.count else { return nil }
-        return playlist[row]
-      }
-      guard let item else { return nil }
-
-      if identifier == .isChosen {
-        let pointer = view.userInterfaceLayoutDirection == .rightToLeft ?
-            Constants.String.blackLeftPointingTriangle :  Constants.String.blackRightPointingTriangle
-        v.textField?.stringValue = item.isPlaying ? pointer : ""
-      } else if identifier == .trackName {
-        let cellView = v as! PlaylistTrackCellView
-        // file name
-        let filename = item.filenameForDisplay
-        let displayStr: String = NSString(string: filename).deletingPathExtension
-
-        func getCachedMetadata() -> (artist: String, title: String)? {
-          guard Preference.bool(for: .playlistShowMetadata) else { return nil }
-          if Preference.bool(for: .playlistShowMetadataInMusicMode) && !player.isInMiniPlayer {
-            return nil
-          }
-          guard let metadata = info.getCachedMetadata(item.filename) else { return nil }
-          guard let artist = metadata.artist, let title = metadata.title else { return nil }
-          return (artist, title)
-        }
-
-        if let prefix = player.info.currentVideosInfo.first(where: { $0.path == item.filename })?.prefix,
-          !prefix.isEmpty,
-          prefix.count <= displayStr.count,  // check whether prefix length > filename length
-          prefix.count >= PrefixMinLength,
-          filename.count > FilenameMinLength {
-          cellView.setPrefix(prefix)
-          cellView.setTitle(String(filename[filename.index(filename.startIndex, offsetBy: prefix.count)...]))
-        } else {
-          cellView.setPrefix(nil)
-          cellView.setTitle(filename)
-        }
-        // playback progress and duration
-        cellView.durationLabel.font = NSFont.monospacedDigitSystemFont(ofSize: NSFont.smallSystemFontSize, weight: .regular)
-        cellView.durationLabel.stringValue = ""
-        player.playlistQueue.async {
-          if let (artist, title) = getCachedMetadata() {
-            DispatchQueue.main.async {
-              cellView.setTitle(title)
-              cellView.setAdditionalInfo(artist)
-            }
-          }
-          if let cached = self.player.info.getCachedVideoDurationAndProgress(item.filename),
-            let duration = cached.duration {
-            // if it's cached
-            if duration > 0 {
-              // if FFmpeg got the duration successfully
-              DispatchQueue.main.async {
-                cellView.durationLabel.stringValue = VideoTime(duration).stringRepresentation
-                if let progress = cached.progress {
-                  cellView.playbackProgressView.percentage = progress / duration
-                  cellView.playbackProgressView.needsDisplay = true
-                }
-              }
-              self.refreshTotalLength()
-            }
-          } else {
-            // get related data and schedule a reload
-            if Preference.bool(for: .prefetchPlaylistVideoDuration) {
-              self.player.refreshCachedVideoInfo(forVideoPath: item.filename)
-              // Only schedule a reload if data was obtained and cached to avoid looping
-              if let cached = self.player.info.getCachedVideoDurationAndProgress(item.filename),
-                  let duration = cached.duration, duration > 0 {
-                // if FFmpeg got the duration successfully
-                self.refreshTotalLength()
-                DispatchQueue.main.async {
-                  self.playlistTableView.reloadData(forRowIndexes: IndexSet(integer: row), columnIndexes: IndexSet(integersIn: 0...1))
-                }
-              }
-            }
-          }
-        }
-        // sub button
-        if !info.isMatchingSubtitles,
-          let matchedSubs = player.info.getMatchedSubs(item.filename), !matchedSubs.isEmpty {
-          cellView.setDisplaySubButton(true)
-        } else {
-          cellView.setDisplaySubButton(false)
-        }
-        // not sure why this line exists, but let's keep it for now
-        cellView.subBtn.image?.isTemplate = true
-      }
-      return v
-    }
-    // chapter
-    else if tableView == chapterTableView {
-      let chapters = info.chapters
-      guard row < chapters.count else {
-        return nil
-      }
-      let chapter = chapters[row]
-      // next chapter time
-      let nextChapterTime = chapters[at: row+1]?.time ?? .infinite
-      // construct view
-
-      if identifier == .isChosen {
-        // left column
-        let pointer = view.userInterfaceLayoutDirection == .rightToLeft ?
-            Constants.String.blackLeftPointingTriangle :  Constants.String.blackRightPointingTriangle
-        v.textField?.stringValue = (info.chapter == row) ? pointer : ""
-        return v
-      } else if identifier == .trackName {
-        // right column
-        let cellView = v as! ChapterTableCellView
-        cellView.setTitle(chapter.title.isEmpty ? "Chapter \(row)" : chapter.title)
-        cellView.durationTextField.stringValue = "\(chapter.time.stringRepresentation) → \(nextChapterTime.stringRepresentation)"
-        return cellView
-      } else {
-        return nil
-      }
-    }
-    else {
-      return nil
-    }
-  }
-
-  // MARK: - Context menu
-
-  var selectedRows: IndexSet?
-
-  func menuNeedsUpdate(_ menu: NSMenu) {
-    let selectedRow = playlistTableView.selectedRowIndexes
-    let clickedRow = playlistTableView.clickedRow
-    var target = IndexSet()
-
-    if clickedRow != -1 {
-      if selectedRow.contains(clickedRow) {
-        target = selectedRow
-      } else {
-        target.insert(clickedRow)
-      }
-    }
-
-    selectedRows = target
-    menu.removeAllItems()
-    let items = buildMenu(forRows: target).items
-    for item in items {
-      menu.addItem(item)
-    }
-  }
-
-  @IBAction func contextMenuPlayNext(_ sender: NSMenuItem) {
-    guard let selectedRows = selectedRows else { return }
-    let current = player.mpv.getInt(MPVProperty.playlistPos)
-    var ob = 0  // index offset before current playing item
-    var mc = 1  // moved item count, +1 because move to next item of current played one
-    for item in selectedRows {
-      if item == current { continue }
-      if item < current {
-        player.playlistMove(item + ob, to: current + mc + ob)
-        ob -= 1
-      } else {
-        player.playlistMove(item, to: current + mc + ob)
-      }
-      mc += 1
-    }
-    playlistTableView.deselectAll(nil)
-    player.postNotification(.iinaPlaylistChanged)
-  }
-
-  @IBAction func contextMenuPlayInNewWindow(_ sender: NSMenuItem) {
-    let files = {
-      self.player.info.$playlist.withLock { playlist in
-        self.selectedRows!.enumerated().map { (_, i) in
-          URL(fileURLWithPath: playlist[i].filename)
-        }}
-    }()
-    PlayerCore.newPlayerCore.openURLs(files, shouldAutoLoad: false)
-  }
-
-  @IBAction func contextMenuRemove(_ sender: NSMenuItem) {
-    guard let selectedRows = selectedRows else { return }
-    player.playlistRemove(selectedRows)
-  }
-
-  @IBAction func contextMenuDeleteFile(_ sender: NSMenuItem) {
-    guard let selectedRows = selectedRows else { return }
-    Logger.log("User chose to delete files from playlist at indexes: \(selectedRows.map{$0})", subsystem: player.subsystem)
-
-    var successes = IndexSet()
-    for index in selectedRows {
-      let playlistItem = player.info.$playlist.withLock { $0[index] }
-      guard !playlistItem.isNetworkResource else { continue }
-      let url = URL(fileURLWithPath: playlistItem.filename)
-      do {
-        Logger.log("Trashing row \(index): \(url.standardizedFileURL)", subsystem: player.subsystem)
-        try FileManager.default.trashItem(at: url, resultingItemURL: nil)
-        successes.insert(index)
-      } catch let error {
-        Utility.showAlert("playlist.error_deleting", arguments: [error.localizedDescription])
-      }
-    }
-    if !successes.isEmpty {
-      player.playlistRemove(successes)
-    }
-  }
-
-  @IBAction func contextMenuDeleteFileAfterPlayback(_ sender: NSMenuItem) {
-    // WIP
-  }
-
-  @IBAction func contextMenuShowInFinder(_ sender: NSMenuItem) {
-    guard let selectedRows = selectedRows else { return }
-    var urls: [URL] = []
-    player.info.$playlist.withLock { playlist in
-      for index in selectedRows {
-        if !playlist[index].isNetworkResource {
-          urls.append(URL(fileURLWithPath: playlist[index].filename))
-        }
-      }
-    }
-    playlistTableView.deselectAll(nil)
-    NSWorkspace.shared.activateFileViewerSelecting(urls)
-  }
-
-  @IBAction func contextMenuAddSubtitle(_ sender: NSMenuItem) {
-    guard let selectedRows = selectedRows, let index = selectedRows.first else { return }
-    let filename = player.info.$playlist.withLock { $0[index].filename }
-    let fileURL = URL(fileURLWithPath: filename).deletingLastPathComponent()
-    Utility.quickMultipleOpenPanel(title: NSLocalizedString("alert.choose_media_file.title", comment: "Choose Media File"), dir: fileURL, canChooseDir: true) { subURLs in
-      for subURL in subURLs {
-        guard Utility.supportedFileExt[.sub]!.contains(subURL.pathExtension.lowercased()) else { return }
-        self.player.info.$matchedSubs.withLock { $0[filename, default: []].append(subURL) }
-      }
-      self.playlistTableView.reloadData(forRowIndexes: selectedRows, columnIndexes: IndexSet(integersIn: 0...1))
-    }
-  }
-
-  @IBAction func contextMenuWrongSubtitle(_ sender: NSMenuItem) {
-    guard let selectedRows = selectedRows else { return }
-    for index in selectedRows {
-      let filename = player.info.$playlist.withLock { $0[index].filename }
-      player.info.$matchedSubs.withLock { $0[filename]?.removeAll() }
-      playlistTableView.reloadData(forRowIndexes: selectedRows, columnIndexes: IndexSet(integersIn: 0...1))
-    }
-  }
-
-  @IBAction func contextOpenInBrowser(_ sender: NSMenuItem) {
-    guard let selectedRows = selectedRows else { return }
-    selectedRows.forEach { i in
-      let info = player.info.playlist[i]
-      if info.isNetworkResource, let url = URL(string: info.filename) {
-        NSWorkspace.shared.open(url)
-      }
-    }
-  }
-
-  @IBAction func contextCopyURL(_ sender: NSMenuItem) {
-    guard let selectedRows = selectedRows else { return }
-    let urls = selectedRows.compactMap { i -> String? in
-      let info = player.info.playlist[i]
-      return info.isNetworkResource ? info.filename : nil
-    }
-    NSPasteboard.general.clearContents()
-    NSPasteboard.general.writeObjects([urls.joined(separator: "\n") as NSString])
-  }
-
-  private func buildMenu(forRows rows: IndexSet) -> NSMenu {
-    let result = NSMenu()
-    let isSingleItem = rows.count == 1
-
-    if !rows.isEmpty {
-      let matchedSubCount = rows.map { index in
-        player.info.$playlist.withLock {
-          player.info.getMatchedSubs($0[index].filename)?.count ?? 0
-        }
-      }.reduce(0, +)
-      let title: String = isSingleItem ?
-        player.info.$playlist.withLock { $0[rows.first!] }.filenameForDisplay:
-        String(format: NSLocalizedString("pl_menu.title_multi", comment: "%d Items"), rows.count)
-
-      result.addItem(withTitle: title)
-      result.addItem(NSMenuItem.separator())
-      if #available(macOS 14.0, *) {
-        result.addItem(.sectionHeader(title: NSLocalizedString("pl_menu.playback", comment: "Playback")))
-      }
-      result.addItem(withTitle: NSLocalizedString("pl_menu.play_next", comment: "Play Next"), action: #selector(self.contextMenuPlayNext(_:)))
-      result.addItem(withTitle: NSLocalizedString("pl_menu.play_in_new_window", comment: "Play in New Window"), action: #selector(self.contextMenuPlayInNewWindow(_:)))
-      result.addItem(withTitle: NSLocalizedString(isSingleItem ? "pl_menu.remove" : "pl_menu.remove_multi", comment: "Remove"), action: #selector(self.contextMenuRemove(_:)))
-
-      if !player.isInMiniPlayer && (isSingleItem || matchedSubCount != 0) {
-        result.addItem(NSMenuItem.separator())
-        if #available(macOS 14.0, *) {
-          result.addItem(.sectionHeader(title: String(format: NSLocalizedString("pl_menu.subtitles", comment: "Subtitles (%@ loaded)"), matchedSubCount)))
-        } else {
-          result.addItem(withTitle: String(format: NSLocalizedString("pl_menu.matched_sub", comment: "Matched %d Subtitle(s)"), matchedSubCount))
-        }
-        if isSingleItem {
-          result.addItem(withTitle: NSLocalizedString("pl_menu.add_sub", comment: "Add Subtitle…"), action: #selector(self.contextMenuAddSubtitle(_:)))
-        }
-        if matchedSubCount != 0 {
-          result.addItem(withTitle: NSLocalizedString("pl_menu.wrong_sub", comment: "Wrong Subtitle"), action: #selector(self.contextMenuWrongSubtitle(_:)))
-        }
-      }
-
-      result.addItem(NSMenuItem.separator())
-      // network resources related operations
-      let networkCount = player.info.$playlist.withLock { playlist in
-        rows.filter { playlist[$0].isNetworkResource }
-      }.count
-      if networkCount != 0 {
-        if #available(macOS 14.0, *) {
-          result.addItem(.sectionHeader(title: NSLocalizedString("pl_menu.network_resources", comment: "Network Resources")))
-        }
-        result.addItem(withTitle: NSLocalizedString("pl_menu.browser", comment: "Open in Browser"), action: #selector(self.contextOpenInBrowser(_:)))
-        result.addItem(withTitle: NSLocalizedString(networkCount == 1 ? "pl_menu.copy_url" : "pl_menu.copy_url_multi", comment: "Copy URL(s)"), action: #selector(self.contextCopyURL(_:)))
-        result.addItem(NSMenuItem.separator())
-      }
-      // file related operations
-      let localCount = rows.count - networkCount
-      if localCount != 0 {
-        if #available(macOS 14.0, *) {
-          result.addItem(.sectionHeader(title: NSLocalizedString("pl_menu.file_operations", comment: "File Operations")))
-        }
-        result.addItem(withTitle: NSLocalizedString(localCount == 1 ? "pl_menu.delete" : "pl_menu.delete_multi", comment: "Delete"), action: #selector(self.contextMenuDeleteFile(_:)))
-        // result.addItem(withTitle: NSLocalizedString(isSingleItem ? "pl_menu.delete_after_play" : "pl_menu.delete_after_play_multi", comment: "Delete After Playback"), action: #selector(self.contextMenuDeleteFileAfterPlayback(_:)))
-
-        result.addItem(withTitle: NSLocalizedString("pl_menu.show_in_finder", comment: "Show in Finder"), action: #selector(self.contextMenuShowInFinder(_:)))
-        result.addItem(NSMenuItem.separator())
-      }
-    }
-
-    // menu items from plugins
-    var hasPluginMenuItems = false
-    let filenames = Array(rows)
-    let pluginMenuItems = player.plugins.map {
-      plugin -> (JavascriptPluginInstance, [JavascriptPluginMenuItem]) in
-      if let builder = (plugin.apis["playlist"] as! JavascriptAPIPlaylist).menuItemBuilder?.value,
-        let value = builder.call(withArguments: [filenames]),
-        value.isObject,
-        let items = value.toObject() as? [JavascriptPluginMenuItem] {
-        hasPluginMenuItems = true
-        return (plugin, items)
-      }
-      return (plugin, [])
-    }
-    if hasPluginMenuItems {
-      result.addItem(withTitle: NSLocalizedString("pl_menu.plugin", comment: "Plugin"))
-      for (plugin, items) in pluginMenuItems {
-        for item in items {
-          add(menuItemDef: item, to: result, for: plugin)
-        }
-      }
-      result.addItem(NSMenuItem.separator())
-    }
-
-    if #available(macOS 14.0, *) {
-      result.addItem(.sectionHeader(title: NSLocalizedString("pl_menu.playlist", comment: "Playlist")))
-    }
-    result.addItem(withTitle: NSLocalizedString("pl_menu.add_file", comment: "Add File"), action: #selector(self.addFileAction(_:)))
-    result.addItem(withTitle: NSLocalizedString("pl_menu.add_url", comment: "Add URL"), action: #selector(self.addURLAction(_:)))
-    result.addItem(withTitle: NSLocalizedString("pl_menu.clear_playlist", comment: "Clear Playlist"), action: #selector(self.clearPlaylistBtnAction(_:)))
-    return result
-  }
-
-  @discardableResult
-  private func add(menuItemDef item: JavascriptPluginMenuItem,
-                   to menu: NSMenu,
-                   for plugin: JavascriptPluginInstance) -> NSMenuItem {
-    if (item.isSeparator) {
-      let item = NSMenuItem.separator()
-      menu.addItem(item)
-      return item
-    }
-
-    let menuItem: NSMenuItem
-    if item.action == nil {
-      menuItem = menu.addItem(withTitle: item.title, action: nil, target: plugin, obj: item)
-    } else {
-      menuItem = menu.addItem(withTitle: item.title,
-                              action: #selector(plugin.playlistMenuItemAction(_:)),
-                              target: plugin,
-                              obj: item)
-    }
-
-    menuItem.isEnabled = item.enabled
-    menuItem.state = item.selected ? .on : .off
-    if !item.items.isEmpty {
-      menuItem.submenu = NSMenu()
-      for submenuItem in item.items {
-        add(menuItemDef: submenuItem, to: menuItem.submenu!, for: plugin)
-      }
-    }
-    return menuItem
-  }
 }
 
-
-class PlaylistTrackCellView: NSTableCellView {
-  @IBOutlet weak var subBtn: NSButton!
-  @IBOutlet weak var subBtnWidthConstraint: NSLayoutConstraint!
-  @IBOutlet weak var subBtnTrailingConstraint: NSLayoutConstraint!
-  @IBOutlet weak var prefixBtn: PlaylistPrefixButton!
-  @IBOutlet weak var infoLabel: NSTextField!
-  @IBOutlet weak var infoLabelTrailingConstraint: NSLayoutConstraint!
-  @IBOutlet weak var durationLabel: NSTextField!
-  @IBOutlet weak var playbackProgressView: PlaylistPlaybackProgressView!
-
-  func setPrefix(_ prefix: String?) {
-    if let prefix = prefix {
-      prefixBtn.hasPrefix = true
-      prefixBtn.text = prefix
-    } else {
-      prefixBtn.hasPrefix = false
-    }
-  }
-
-  func setDisplaySubButton(_ show: Bool) {
-    if show {
-      subBtn.isHidden = false
-      subBtnWidthConstraint.constant = 12
-      subBtnTrailingConstraint.constant = 4
-    } else {
-      subBtn.isHidden = true
-      subBtnWidthConstraint.constant = 0
-      subBtnTrailingConstraint.constant = 0
-    }
-  }
-
-  func setAdditionalInfo(_ string: String?) {
-    if let string = string {
-      infoLabel.isHidden = false
-      infoLabelTrailingConstraint.constant = 4
-      infoLabel.stringValue = string
-      infoLabel.toolTip = string
-    } else {
-      infoLabel.isHidden = true
-      infoLabelTrailingConstraint.constant = 0
-      infoLabel.stringValue = ""
-    }
-  }
-
-  func setTitle(_ title: String) {
-    textField?.stringValue = title
-    textField?.toolTip = title
-  }
-
-  override func prepareForReuse() {
-    super.prepareForReuse()
-    playbackProgressView.percentage = 0
-    playbackProgressView.needsDisplay = true
-    setPrefix(nil)
-    setAdditionalInfo(nil)
-  }
-}
-
-
-class PlaylistPrefixButton: NSButton {
-
-  var text = "" {
-    didSet {
-      refresh()
-    }
-  }
-
-  var hasPrefix = true {
-    didSet {
-      refresh()
-    }
-  }
-
-  var isFolded = true {
-    didSet {
-      refresh()
-    }
-  }
-
-  private func refresh() {
-    self.title = hasPrefix ? (isFolded ? "…" : text) : ""
-  }
-
-}
-
-class SubPopoverViewController: NSViewController, NSTableViewDelegate, NSTableViewDataSource {
-
-  @IBOutlet weak var tableView: NSTableView!
-  @IBOutlet weak var playlistTableView: NSTableView!
-  @IBOutlet weak var heightConstraint: NSLayoutConstraint!
-
-  weak var player: PlayerCore!
-
-  var filePath: String = ""
-
-  func tableView(_ tableView: NSTableView, shouldSelectRow row: Int) -> Bool {
-    return false
-  }
-
-  func tableView(_ tableView: NSTableView, objectValueFor tableColumn: NSTableColumn?, row: Int) -> Any? {
-    guard let matchedSubs = player.info.getMatchedSubs(filePath) else { return nil }
-    return matchedSubs[row].lastPathComponent
-  }
-
-  func numberOfRows(in tableView: NSTableView) -> Int {
-    return player.info.getMatchedSubs(filePath)?.count ?? 0
-  }
-
-  @IBAction func wrongSubBtnAction(_ sender: AnyObject) {
-    player.info.$matchedSubs.withLock { $0[filePath]?.removeAll() }
-    tableView.reloadData()
-    if let row = player.info.$playlist.withLock({ $0.firstIndex(where: { $0.filename == filePath }) }) {
-      playlistTableView.reloadData(forRowIndexes: IndexSet(integer: row), columnIndexes: IndexSet(integersIn: 0...1))
-    }
-  }
-}
-
-class ChapterTableCellView: NSTableCellView {
-  @IBOutlet weak var durationTextField: NSTextField!
-
-  func setTitle(_ title: String) {
-    textField?.stringValue = title
-    textField?.toolTip = title
-  }
-}
