+++ iina/SidebarPane.swift
@@ -1,5 +1,5 @@
 //
-//  AnimatedTabViewController.swift
+//  SidebarPane.swift
 //  iina
 //
 //  Created by Hechen Li on 2026-05-30.
@@ -7,11 +7,12 @@
 //
 
 
-class AnimatedTabViewController: NSTabViewController {
+class SidebarTabViewController: NSTabViewController {
   let transitionDuration: TimeInterval = 0.3
   private let prefObserver = Preference.Observer()
 
-  private var previousIndex: Int = -1
+  // -1 for the first tab switch, don't show animation in this case
+  var previousIndex: Int = -1
 
   override func viewDidLoad() {
     super.viewDidLoad()
@@ -37,7 +38,7 @@ class AnimatedTabViewController: NSTabViewController {
   ) {
     if Preference.bool(for: .disableAnimations) || previousIndex < 0 {
       previousIndex = selectedTabViewItemIndex
-      super.transition(from: fromVC, to: toVC, options: options, completionHandler: completion)
+      super.transition(from: fromVC, to: toVC, options: [], completionHandler: completion)
       return
     }
 
@@ -65,7 +66,11 @@ class AnimatedTabViewController: NSTabViewController {
 }
 
 
-class SidebarScrollView: NSScrollView {
+protocol SidebarPane: NSView {
+  var horizontalScroll: ((Bool) -> Void)? { get set }
+}
+
+class SidebarScrollView: NSScrollView, SidebarPane {
   class Container: NSBox {
     init(_ view: NSView, _ block: (NSView) -> Void) {
       super.init(frame: .zero)
