+++ iina/en.lproj/Localizable.strings
@@ -541,6 +541,14 @@
 
 // Sidebar
 
+"sidebar.layout" = "Layout";
+"sidebar.video" = "Video";
+"sidebar.audio" = "Audio";
+"sidebar.sub" = "Subtitles";
+"sidebar.plugins" = "Plugins";
+"sidebar.playlist" = "Playlist";
+"sidebar.chapters" = "Chapters";
+
 "sidebar.aspect_ratio" = "Aspect Ratio";
 "sidebar.crop" = "Crop";
 "sidebar.rotation" = "Rotation";
@@ -575,3 +583,6 @@
 "sidebar.background" = "Background";
 "sidebar.sub_settings_not_available" = "Some settings are not available for this subtitle type.";
 "sidebar.sub_style_warning" = "Subtitle style options may break ASS rendering, and some of them will be disabled depending on subtitle type.";
+
+"sidebar.add_file" = "Add File…";
+"sidebar.add_url" = "Add URL…";
